+++ doc/source/history.rst
@@ -7,9 +7,10 @@ History
 
     - This a LTS release: it will be updated with bug fixes, without
       breaking compatibility.
-    - new ``Client::wait`` performs a ``Client::pull`` after waiting for a new
-      push from another client. This allows getting new data as soon as it is
-      available, without having to poll.
+    - ``Client::pull`` now takes a ``wait_milliseconds`` optional parameter: if
+      the server has no data to be pulled, it will wait for new data for this
+      duration. This allows getting new data as soon as it is available,
+      without having to poll.
     - ``Encoded_File`` supports on-the-fly coding or decoding of data.
       ``Brotli_Codec`` is provided for compression.
     - It is now possible to write to an ``Interpreted_File``.